Move component to its own directory
