+++ packages/js/components/changelog/add-tree-select
@@ -1,0 +1,4 @@
+Significance: minor
+Type: add
+
+Create SelectTree component that uses TreeControl
