+++ packages/js/components/src/experimental-select-tree-control/index.ts
@@ -1,0 +1,1 @@
+export * from './select-tree';
