+++ packages/js/components/src/experimental-tree-control/index.ts
@@ -2,4 +2,3 @@ export * from './tree';
 export * from './tree-control';
 export * from './tree-item';
 export * from './types';
-export * from './select-tree';
