+++ packages/js/components/src/index.ts
@@ -98,8 +98,8 @@ export {
 export {
 	TreeControl as __experimentalTreeControl,
 	Item as TreeItemType,
-	SelectTree as __experimentalSelectTreeControl,
 } from './experimental-tree-control';
+export { SelectTree as __experimentalSelectTreeControl } from './experimental-select-tree-control';
 export { default as TreeSelectControl } from './tree-select-control';
 
 // Exports below can be removed once the @woocommerce/product-editor package is released.
